requirements.txt
ccxt
pandas
numpy
matplotlib
python-dateutil
requests
